+++ HISTORY.md
@@ -7,6 +7,7 @@ https://github.com/josdejong/mathjs
 - Renamed unit conversion function and operator `in` to `to`. Operator `in` is
   still available in the expression parser as an alias for `to`. Added unit
   `in`, an abbreviation for `inch`. Thanks Elijah Insua (tmpvar).
+- Implemented an argument `includeEnd` for function `range` (false by default).
 
 
 ## 2013-12-19, version 0.17.1